tests: refactor break up tests and remove rarely used types
