+++ contracts/sdjwt-verifier/src/tests/mod.rs
@@ -1,3 +1,6 @@
-mod contract_test;
+mod deregister;
 mod fixtures;
-mod verifier_test;
+mod register;
+mod revocation;
+mod update;
+mod verifier;
